@@ -10,13 +10,24 @@ import com.lms.course_service.entities.ModuleEntity;
 import com.lms.course_service.mappers.ModuleMapper;
 import com.lms.course_service.repositories.CourseRepository;
 import com.lms.course_service.repositories.ModuleRepository;
+import com.lms.shared.events.CheckResourceExistsEvent;
+import com.lms.shared.events.ResourceExistsResponseEvent;
 import lombok.AllArgsConstructor;
+import org.apache.kafka.clients.consumer.ConsumerRecord;
+import org.apache.kafka.clients.producer.ProducerRecord;
+import org.apache.kafka.common.header.internals.RecordHeader;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.http.HttpStatus;
+import org.springframework.kafka.core.KafkaTemplate;
+import org.springframework.kafka.requestreply.ReplyingKafkaTemplate;
+import org.springframework.kafka.requestreply.RequestReplyFuture;
+import org.springframework.kafka.support.KafkaHeaders;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.UUID;
+import java.util.concurrent.ExecutionException;
 import java.util.stream.Collectors;
 
 @Service
@@ -25,21 +36,20 @@ public class ModuleService {
 
     private final ModuleRepository moduleRepository;
     private final CourseRepository courseRepository;
+    private final ReplyingKafkaTemplate<String, Object, ResourceExistsResponseEvent> resourceReplyingKafkaTemplate;
 
     public ServiceResult<ModuleResponse> createModule(CreateModuleRequest request) {
-
         Optional<CourseEntity> courseOptional = courseRepository.findCourseEntitiesById(request.getCourseId());
-        CourseEntity course = null;
 
-        if (courseOptional.isPresent()) {
-            course = courseOptional.get();
-        } else {
+        if (courseOptional.isEmpty()) {
             return ServiceResult.<ModuleResponse>builder()
                     .success(false)
                     .httpStatus(HttpStatus.NOT_FOUND)
                     .messageError("Course not found for ID: " + request.getCourseId())
                     .build();
         }
+
+        CourseEntity course = courseOptional.get();
 
         ModuleEntity moduleEntity = ModuleEntity.builder()
                 .title(request.getTitle())
@@ -58,19 +68,18 @@ public class ModuleService {
     }
 
     public List<ModuleResponse> getModules(ModuleFilterParams filterParams) {
-
         List<ModuleEntity> moduleEntities = moduleRepository.findModulesEntitiesByFilters(
                 filterParams.getModuleId(),
                 filterParams.getCourseID(),
                 PageRequest.of(filterParams.getPage(), filterParams.getSize())
         ).stream().toList();
 
-
-        return moduleEntities.stream().map(ModuleMapper::toResponse).collect(Collectors.toList());
+        return moduleEntities.stream()
+                .map(ModuleMapper::toResponse)
+                .collect(Collectors.toList());
     }
 
-    public ServiceResult<ModuleResponse> addResourceToModule(AddResourceToModuleRequest request ) {
-
+    public ServiceResult<ModuleResponse> addResourceToModule(AddResourceToModuleRequest request) {
         if (!moduleRepository.existsById(request.getModuleId())) {
             return ServiceResult.<ModuleResponse>builder()
                     .success(false)
@@ -79,17 +88,61 @@ public class ModuleService {
                     .build();
         }
 
-        //todo: check if the resource exists in the database
+        try {
+            String correlationId = UUID.randomUUID().toString();
+            CheckResourceExistsEvent event = new CheckResourceExistsEvent(List.of(request.getResourceId()), correlationId);
 
-        ModuleEntity moduleEntity = moduleRepository.findModuleEntityById(request.getModuleId());
+            ProducerRecord<String, Object> record = new ProducerRecord<>("resource-validation-request", event);
+            record.headers().add(new RecordHeader(KafkaHeaders.REPLY_TOPIC, "resource-validation-response".getBytes()));
+            record.headers().add(new RecordHeader(KafkaHeaders.CORRELATION_ID, correlationId.getBytes()));
 
-        moduleEntity.getResourceIds().add(request.getResourceId());
-        moduleRepository.save(moduleEntity);
+            RequestReplyFuture<String, Object, ResourceExistsResponseEvent> future =
+                    resourceReplyingKafkaTemplate.sendAndReceive(record);
 
-        return ServiceResult.<ModuleResponse>builder()
-                .data(ModuleMapper.toResponse(moduleEntity))
-                .success(true)
-                .httpStatus(HttpStatus.OK)
-                .build();
+            ConsumerRecord<String, ResourceExistsResponseEvent> response = future.get();
+
+            if (response == null || response.value() == null ||
+                    !response.value().validResourceIds().contains(request.getResourceId())) {
+                return ServiceResult.<ModuleResponse>builder()
+                        .success(false)
+                        .httpStatus(HttpStatus.BAD_REQUEST)
+                        .messageError("Invalid resource ID: " + request.getResourceId())
+                        .build();
+            }
+
+
+            ModuleEntity moduleEntity = moduleRepository.findModuleEntityById(request.getModuleId());
+
+            if (moduleEntity != null && moduleEntity.getResourceIds().contains(request.getResourceId())) {
+                return ServiceResult.<ModuleResponse>builder()
+                        .success(false)
+                        .httpStatus(HttpStatus.BAD_REQUEST)
+                        .messageError("The resource ID: " + request.getResourceId() + " already exists!")
+                        .build();
+            }
+
+            moduleEntity.getResourceIds().add(request.getResourceId());
+            moduleRepository.save(moduleEntity);
+
+            return ServiceResult.<ModuleResponse>builder()
+                    .data(ModuleMapper.toResponse(moduleEntity))
+                    .success(true)
+                    .httpStatus(HttpStatus.OK)
+                    .build();
+
+        } catch (InterruptedException e) {
+            Thread.currentThread().interrupt();
+            return ServiceResult.<ModuleResponse>builder()
+                    .success(false)
+                    .httpStatus(HttpStatus.INTERNAL_SERVER_ERROR)
+                    .messageError("Kafka interrupted: " + e.getMessage())
+                    .build();
+        } catch (ExecutionException e) {
+            return ServiceResult.<ModuleResponse>builder()
+                    .success(false)
+                    .httpStatus(HttpStatus.INTERNAL_SERVER_ERROR)
+                    .messageError("Kafka error: " + e.getMessage())
+                    .build();
+        }
     }
 }
