@@ -11,22 +11,20 @@ import org.springframework.context.annotation.Configuration;
 public class SwaggerConfig {
     @Bean
     public OpenAPI customOpenAPI() {
-        // Define the security scheme
         SecurityScheme securityScheme = new SecurityScheme()
                 .type(SecurityScheme.Type.HTTP)
                 .scheme("bearer")
                 .bearerFormat("JWT")
                 .description("Enter your Bearer token to access secured endpoints");
 
-        // Add security requirements to the OpenAPI object
         SecurityRequirement securityRequirement = new SecurityRequirement()
                 .addList("BearerAuth");
 
         return new OpenAPI()
                 .info(new Info()
-                        .title("LMS Backend API")
+                        .title("Assignment Service of LMS Backend API")
                         .version("1.0")
-                        .description("API documentation for the Learning Management System backend"))
+                        .description("API documentation for Assignment Service of the Learning Management System backend"))
                 .addSecurityItem(securityRequirement)
                 .components(new io.swagger.v3.oas.models.Components()
                         .addSecuritySchemes("BearerAuth", securityScheme));
