@@ -2,9 +2,16 @@ package com.lms.assignment_service.mappers;
 
 import com.lms.assignment_service.dtos.responses.AssignmentSubmissionResponse;
 import com.lms.assignment_service.entities.AssignmentSubmissionEntity;
+import com.lms.assignment_service.kafka.ResourceRequestDispatcher;
+import com.lms.shared.dtos.ResourceResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
 
 import java.util.Collections;
+import java.util.List;
 
+@Component
+@RequiredArgsConstructor
 public class AssignmentSubmissionMapper {
     public static AssignmentSubmissionResponse toResponse(AssignmentSubmissionEntity entity) {
         if (entity == null) {return null;}
@@ -17,7 +24,7 @@ public class AssignmentSubmissionMapper {
                 .grade(entity.getGrade())
                 .completed(entity.getCompleted())
                 .comment(entity.getComment())
-                .resources(Collections.emptyList()) // todo: get the list of the resources response from the resource service
+                .resources(entity.getResourceIds())
                 .build();
     }
 }
