@@ -24,9 +24,9 @@ public class SwaggerConfig {
 
         return new OpenAPI()
                 .info(new Info()
-                        .title("LMS Backend API")
+                        .title("Authentication Service of LMS Backend API")
                         .version("1.0")
-                        .description("API documentation for the Learning Management System backend"))
+                        .description("API documentation for Authentication Service of the Learning Management System backend"))
                 .addSecurityItem(securityRequirement)
                 .components(new io.swagger.v3.oas.models.Components()
                         .addSecuritySchemes("BearerAuth", securityScheme));
