@@ -3,17 +3,19 @@ package com.lms.auth_service.entities.relational;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 
 import java.util.List;
 
-@Entity
-@Table(name = "admins")
+@EqualsAndHashCode(callSuper = true)
 @AllArgsConstructor
 @NoArgsConstructor
 @SuperBuilder
 @Data
+@Entity
+@Table(name = "admins")
 public class AdminEntity extends UserEntity {
 
     @Column(nullable = false)
