@@ -4,7 +4,10 @@ import jakarta.persistence.*;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
 
-
+@EqualsAndHashCode(callSuper = true)
+@Data
+@NoArgsConstructor
+@SuperBuilder
 @Entity
 @Table(name = "students")
 public class StudentEntity extends UserEntity {
