@@ -22,7 +22,7 @@ public class JwtUtil {
             final String extractedRole = extractRole(token);
             return extractedEmail.equals(email) && extractedRole.equals(requiredRole) && !isTokenExpired(token);
         } catch (Exception e) {
-            return false; // Invalid token
+            return false;
         }
     }
 
